+++ stylelib/masks.h
@@ -13,7 +13,7 @@ namespace DSP
 class Q_DECL_EXPORT Mask
 {
 public:
-    static void render(const QRectF &r, const QBrush &b, QPainter *p, const quint8 round, const Sides s = All, const QPoint &offset = QPoint());
+    static void render(const QRectF &r, const QBrush &b, QPainter *p, const float round, const Sides s = All, const QPoint &offset = QPoint());
 };
 
 }
